Rename dataset modules to snake_case and update imports
