+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 python_files = *.py
-addopts = --ignore=src/StreamlitUI.py
+addopts = --ignore=src/streamlit_ui_legacy.py
           --ignore=src/APIExplorerUI.py
