+++ src/langunits/LangUnit.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod, ABCMeta
 from enum import Enum
 from dataclasses import dataclass
-from data.Dataset import Unit
+from data.dataset import Unit
 
 
 class UnitType(Enum):
