+++ src/langunits/RegexVal/RegexVal.py
@@ -1,7 +1,7 @@
 from colorama import Fore
 from deprecated import deprecated
 
-from data.Dataset import Unit
+from data.dataset import Unit
 from langunits.LangUnit import LangUnit, UnitType, EvalRequest, EvalResponse
 import re
