+++ src/langunits/SqlSelect/SqlSelect.py
@@ -2,7 +2,7 @@ from typing import List
 
 from deprecated import deprecated
 
-from data.Dataset import Unit
+from data.dataset import Unit
 from langunits.LangUnit import LangUnit, UnitType, EvalRequest, EvalResponse
