+++ src/models/providers/GroqModelProvider.py
@@ -3,7 +3,7 @@ import os
 from langunits.LangUnit import LangUnitInfo
 from models.ModelBase import ModelBase
 from colorama import init, Fore, Back, Style
-from data.Dataset import *
+from data.dataset import *
 from models.ModelBase import GenResponse, GenRequest
 from models.providers.ModelProviderBase import ModelProviderBase
