+++ src/streamlit_ui.py
@@ -8,7 +8,7 @@ import pandas as pd
 import streamlit as st
 
 from api.api import API
-from data.Dataset import Dataset, Unit
+from data.dataset import Dataset, Unit
 from experiments.ModelConfiguration import ModelConfiguration, StaticModelConfiguration
 from langunits.LangUnitFactory import LangUnitFactory
 from models.ModelBase import ModelProviderMeta
